requirements.txt
cffi==1.14.5
cryptg==0.2.post4
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
rsa==4.7.2
Telethon==1.23.0
requests==2.26.0
python-dotenv==0.19.1
